Teach DNS parser about specific mDNS class bits
This adds support for two mDNS special cases.

The top bits of both Class and QueryClass have special meaning.

For query classes, the top bit indicates whether or not unicast
responses are preferred.

For resource record classes, the top bit indicates whether or not there
are more resource records in future packets.